restructure component test package
