+++ src/integration-test/groovy/org/pih/warehouse/common/base/IntegrationSpec.groovy
@@ -1,0 +1,13 @@
+package org.pih.warehouse.common.base
+
+import grails.test.mixin.integration.Integration
+import org.springframework.context.annotation.Import
+import spock.lang.Specification
+
+/**
+ * Base class for all integration tests.
+ */
+@Integration
+@Import(IntegrationSpecConfig.class)
+abstract class IntegrationSpec extends Specification {
+}
