--- src/integration-test/groovy/org/pih/warehouse/component/api/base/UnauthenticatedApi.groovy
@@ -1,19 +1,0 @@
-package org.pih.warehouse.component.api.base
-
-import io.restassured.specification.RequestSpecification
-
-/**
- * Base class for all API classes that do not require an authenticated user.
- */
-abstract class UnauthenticatedApi extends Api {
-
-    final UnauthenticatedApiContext unauthenticatedApiContext
-
-    UnauthenticatedApi(UnauthenticatedApiContext unauthenticatedApiContext) {
-        this.unauthenticatedApiContext = unauthenticatedApiContext
-    }
-
-    RequestSpecification getBaseRequestSpec() {
-        return unauthenticatedApiContext.baseRequestSpec
-    }
-}
