+++ src/integration-test/groovy/org/pih/warehouse/smoke/spec/base/SmokeSpec.groovy
@@ -1,0 +1,10 @@
+package org.pih.warehouse.smoke.spec.base
+
+import org.pih.warehouse.common.base.IntegrationSpec
+
+/**
+ * Base class for all smoke tests.
+ */
+class SmokeSpec extends IntegrationSpec {
+
+}
